refactor: Add action dispatch, delete view extensions, slim view (#288)
