--- minimark/Views/Window/Flow/ReaderWindowRootView+WarningFlow.swift
@@ -1,3 +1,0 @@
-// Methods moved to ReaderWindowCoordinator (Task 6, #288).
-// This file will be deleted in Task 7.
-import Foundation
